+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Warns when trying to publish a telemetry to both 'http' and 'text' media
     - services.pl by-property new verb, thus bumping minor candidate version number
     - Deprecate 'ttp.pl writejson' in favor of 'ttp.pl writefile' (todo #61)
+    - TTP::commandExec() now honors <ENVIRONMENT> macro (todo #177)
 
 ### 4.31.0